Cast scaling result to int for faster comparison.

+++ compiler/src/org.graalvm.compiler.truffle.runtime/src/org/graalvm/compiler/truffle/runtime/OptimizedCallTarget.java
@@ -561,7 +561,7 @@ public abstract class OptimizedCallTarget implements CompilableTruffleAST, RootC
                          */
                         && !(getRootNode() instanceof OSRRootNode) //
                         && intCallCount >= engine.callThresholdInInterpreter //
-                        && intLoopCallCount >= engine.callAndLoopThresholdInInterpreter * runtime().compilationThresholdScale(); //
+                        && intLoopCallCount >= (int) (engine.callAndLoopThresholdInInterpreter * runtime().compilationThresholdScale()); //
     }
 
     public final boolean shouldCompile() {
@@ -596,7 +596,7 @@ public abstract class OptimizedCallTarget implements CompilableTruffleAST, RootC
         if (!compilationFailed //
                         && !isSubmittedForCompilation()//
                         && firstTierCallCount >= engine.callThresholdInFirstTier //
-                        && firstTierLoopCallCount >= engine.callAndLoopThresholdInFirstTier * runtime().compilationThresholdScale()) {
+                        && firstTierLoopCallCount >= (int) (engine.callAndLoopThresholdInFirstTier * runtime().compilationThresholdScale())) {
             return lastTierCompile();
         }
         return false;
